Remove endpoint to update a submission
A flattened resource of the form data should now also be created on save
of a concept submission, not only on submit. As a consequence the form
needs to be filled in in the backend to retrieve the form data. Since
the validate-submission-service already contains the logic to fill in
a form in the backend, the update endpoint has moved to this service.