Use form file linked to the submission for loading
We use to use the active form file to load concept forms, causing issues
in the case where the forms configuration changed between the moment the
concept form was created and the moment we tried to update it. We now
use the form configuration at creation time to avoid conflicting
behaviours.