Rewrite of the enricher
This seems scary, but it does not have to be. What I did is to remove
the batching process on adding triples from a where clause of a SPARQL
query. This mechanism was too brittle and was not even used in all cases,
especially not in the cases where it was necessary. The batching process
does not even provide a benefit because the triplestore can handle large
responses. The bottleneck would be RAM usage while ingesting all those
results.
All that being removed, the code is much more concise and just as clear
to me (or even clearer).