Relax graph specification for reads
When the triplestore is read (SELECT only), we can relax the strictness
of the graph a little. The organisationId needs to be part of the
graph's URI instead of the URI being a specific string. This is useful
in applications where they use this service for reading submissions, but
the URIs of the graphs are built differently.